requirements.txt
torch
torchvision
numpy
scikit-learn
tqdm
Pillow
tensorboard
pandas
matplotlib
psutil
